+++ .eslintrc.js
@@ -1,1 +1,1 @@
-module.exports = { extends: "@react-native-community" };
+module.exports = {extends: ['@react-native-community']};
