+++ .flowconfig
@@ -37,6 +37,7 @@ esproposal.nullish_coalescing=enable
 module.file_ext=.js
 module.file_ext=.json
 module.file_ext=.ios.js
+module.file_ext=.android.js
 
 munge_underscores=true
