+++ package.json
@@ -28,8 +28,6 @@
     "android",
     "ios",
     "js",
-    "index.js",
-    "index.d.ts",
     "react-native-progress-bar-android.podspec"
   ],
   "author": "Wagner Maciel <wagnermaciel@google.com>",
@@ -47,7 +45,7 @@
     "react-timer-mixin": "^0.13.4",
     "@babel/core": "^7.10.3",
     "@babel/runtime": "^7.10.3",
-    "@react-native-community/eslint-config": "^0.0.7",
+    "@react-native-community/eslint-config": "^1.0.0",
     "@types/react-native": "^0.62.2",
     "babel-jest": "^25.1.0",
     "babel-plugin-module-resolver": "^3.2.0",
